+++ lib/core/src/display/shape/primitive/def/sdf.rs
@@ -13,7 +13,6 @@ use crate::display::shape::primitive::def::class::ShapeRef;
 use crate::display::shape::primitive::shader::canvas::Canvas;
 use crate::display::shape::primitive::shader::canvas::CanvasShape;
 use crate::display::shape::primitive::shader::data::ShaderData;
-use crate::system::gpu::data::BufferItem;
 use crate::display::render::webgl::glsl::Glsl;
 
 use crate::display::render::webgl::glsl::traits::*;
