+++ lib/core/src/system/gpu.rs
@@ -1,4 +1,3 @@
 //! GPU-specific implementations.
 
-pub mod buffer;
 pub mod data;
