+++ lib/core/src/system/gpu/data/attribute.rs
@@ -7,9 +7,9 @@ use crate::data::dirty::traits::*;
 use crate::data::dirty;
 use crate::debug::stats::Stats;
 use crate::display::render::webgl::Context;
-use crate::system::gpu::buffer::IsBuffer;
+use crate::system::gpu::data::buffer::IsBuffer;
 use crate::system::gpu::data::BufferItem;
-use crate::system::gpu::buffer;
+use crate::system::gpu::data::buffer;
 use crate::system::web::group;
 use crate::system::web::Logger;
 use data::opt_vec::OptVec;
